+++ src/controllers/ticketController.js
@@ -5,6 +5,10 @@ class TicketController {
     async createTicket(purchaser, amount, cart) {
         return await ticketService.createTicket(purchaser, amount, cart);
     }
+
+    async getTicketByCode(ticket) {
+        return await ticketService.getTicketByCode(ticket);
+    }
 }
 
 export default TicketController;